requirements.txt
pypdf
python-pptx
openpyxl
